release 0.17.1 to testpypi
git-svn-id: http://svn.code.sf.net/p/docutils/code/trunk@8707 929543f6-e4f2-0310-98a6-ba3bd3dd1d04

+++ sandbox/infrastructure/crosstest/t.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 
-# test 
+# test local wheel
 WHL=../docutils-0.17.1b2.dev0-py2.py3-none-any.whl
+WHL=TESTPYPI
+# PRE =
 
 for VDIR in ./py3.* ; do
     echo "$VDIR"
@@ -11,7 +13,14 @@ for VDIR in ./py3.* ; do
     export PYTHONPATH= ; . bin/activate
     export PYTHONWARNINGS=default
     export LC_ALL=C
-    python -m pip install $WHL
+    if [ -e $WHL ] ; then
+        python -m pip install $WHL
+    elif [ $WHL = "TESTPYPI" ] ; then
+        python -m pip install --no-cache-dir --index-url https://test.pypi.org/simple/ --no-deps docutils
+        python -m pip install --upgrade --index-url https://test.pypi.org/simple/ --no-deps docutils
+    else
+        python -m pip install $WHL
+    fi
     cp -Lr ~/projects/docutils-code/docutils/test . 
     python test/alltests.py
     read X
